+++ native/cocos/renderer/gfx-base/GFXRenderPass.h
@@ -36,7 +36,7 @@ public:
     RenderPass();
     ~RenderPass() override;
 
-    static size_t computeHash(const RenderPassInfo &info);
+    static uint32_t computeHash(const RenderPassInfo &info);
 
     void initialize(const RenderPassInfo &info);
     void destroy();
@@ -45,10 +45,10 @@ public:
     inline const DepthStencilAttachment &getDepthStencilAttachment() const { return _depthStencilAttachment; }
     inline const SubpassInfoList &       getSubpasses() const { return _subpasses; }
     inline const SubpassDependencyList & getDependencies() const { return _dependencies; }
-    inline size_t                        getHash() const { return _hash; }
+    inline uint32_t                        getHash() const { return _hash; }
 
 protected:
-    size_t computeHash();
+    uint32_t computeHash();
 
     virtual void doInit(const RenderPassInfo &info) = 0;
     virtual void doDestroy()                        = 0;
@@ -57,7 +57,7 @@ protected:
     DepthStencilAttachment _depthStencilAttachment;
     SubpassInfoList        _subpasses;
     SubpassDependencyList  _dependencies;
-    size_t                 _hash = 0;
+    uint32_t               _hash = 0;
 };
 
 } // namespace gfx
