+++ native/cocos/renderer/gfx-gles2/GLES2GPUObjects.h
@@ -252,7 +252,7 @@ struct GLES2GPUInputAssembler {
     GLES2GPUBuffer *                     gpuIndirectBuffer = nullptr;
     GLES2GPUAttributeList                glAttribs;
     GLenum                               glIndexType = 0;
-    ccstd::unordered_map<size_t, GLuint> glVAOs;
+    ccstd::unordered_map<uint32_t, GLuint> glVAOs;
 };
 
 struct GLES2GPURenderPass {
